Added packets for party

+++ src/rosecommon/include/components/advancedinfo.h
@@ -11,8 +11,8 @@ struct AdvancedInfo {
     AdvancedInfo() : zuly_(0), hp_(0), mp_(0), moveMode_(AdvancedInfo::RUN), runSpeed_(0), atkSpeed_(0), weightRate_(0) {}
 
     uint64_t zuly_;
-    uint32_t hp_;
-    uint32_t mp_;
+    int16_t hp_;
+    int16_t mp_;
     uint8_t moveMode_;
     uint16_t runSpeed_;
     uint16_t atkSpeed_;
